[lg-1] Application was splitted to modules. Added file appender to logback.
