+++ movieland-persistence-api/src/main/java/com/ivandanylov/movieland/persistence/api/MovieDao.java
@@ -1,0 +1,11 @@
+package com.ivandanylov.movieland.persistence.api;
+
+import com.ivandanylov.movieland.entity.Movie;
+
+import java.util.List;
+
+public interface MovieDao {
+    List<Movie> getAll();
+
+    List<Movie> getRandom();
+}
