+++ movieland-persistence-jdbc/src/main/java/com/ivandanylov/movieland/persistence/jdbc/datasource/HKDataSource.java
@@ -1,0 +1,20 @@
+package com.ivandanylov.movieland.persistence.jdbc.datasource;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+
+import javax.sql.DataSource;
+
+public class HKDataSource {
+    private static HikariConfig config = new HikariConfig();
+
+    private HKDataSource() {
+    }
+
+    public static DataSource getDataSource(String driverClassName, String jdbcUrl) {
+        config.setDriverClassName(driverClassName);
+        config.setJdbcUrl(jdbcUrl);
+
+        return new HikariDataSource(config);
+    }
+}
