+++ movieland-persistence-jdbc/src/test/java/com/ivandanylov/movieland/persistence/jdbc/TestPersistenceJdbcConfig.java
@@ -1,0 +1,22 @@
+package com.ivandanylov.movieland.persistence.jdbc;
+
+import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.stereotype.Repository;
+
+@Configuration
+public class TestPersistenceJdbcConfig {
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer hsqldbProperties() {
+        PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer = new PropertySourcesPlaceholderConfigurer();
+        YamlPropertiesFactoryBean yaml = new YamlPropertiesFactoryBean();
+        yaml.setResources(new ClassPathResource("hsqldb.yml"));
+        propertySourcesPlaceholderConfigurer.setProperties(yaml.getObject());
+        return propertySourcesPlaceholderConfigurer;
+    }
+}
