+++ movieland-service-api/src/main/java/com/ivandanylov/movieland/service/api/MovieService.java
@@ -1,0 +1,11 @@
+package com.ivandanylov.movieland.service.api;
+
+import com.ivandanylov.movieland.entity.Movie;
+
+import java.util.List;
+
+public interface MovieService {
+    List<Movie> getAll();
+
+    List<Movie> getRandom();
+}
