+++ movieland-web/src/main/java/com/ivandanylov/movieland/web/config/DispatcherConfig.java
@@ -1,0 +1,11 @@
+package com.ivandanylov.movieland.web.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = "com.ivandanylov.movieland.web.controller")
+public class DispatcherConfig {
+}
