--- src/main/java/com/ivandanylov/movieland/config/db/HKDataSource.java
@@ -1,20 +1,0 @@
-package com.ivandanylov.movieland.config.db;
-
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-
-import javax.sql.DataSource;
-
-public class HKDataSource {
-    private static HikariConfig config = new HikariConfig();
-
-    private HKDataSource() {
-    }
-
-    public static DataSource getDataSource(String driverClassName, String jdbcUrl) {
-        config.setDriverClassName(driverClassName);
-        config.setJdbcUrl(jdbcUrl);
-
-        return new HikariDataSource(config);
-    }
-}
