--- src/main/java/com/ivandanylov/movieland/config/yaml/YamlApplication.java
@@ -1,20 +1,0 @@
-package com.ivandanylov.movieland.config.yaml;
-
-public class YamlApplication {
-    private YamlDatabase application;
-
-    public YamlDatabase getApplication() {
-        return application;
-    }
-
-    public void setApplication(YamlDatabase application) {
-        this.application = application;
-    }
-
-    @Override
-    public String toString() {
-        return "YamlApplication{" +
-                "application=" + application +
-                '}';
-    }
-}
