--- src/main/java/com/ivandanylov/movieland/config/yaml/YamlDatabase.java
@@ -1,20 +1,0 @@
-package com.ivandanylov.movieland.config.yaml;
-
-public class YamlDatabase {
-    private YamlProperty database;
-
-    public YamlProperty getDatabase() {
-        return database;
-    }
-
-    public void setDatabase(YamlProperty database) {
-        this.database = database;
-    }
-
-    @Override
-    public String toString() {
-        return "YamlDatabase{" +
-                "database=" + database +
-                '}';
-    }
-}
