--- src/main/java/com/ivandanylov/movieland/controller/MovieController.java
@@ -1,34 +1,0 @@
-package com.ivandanylov.movieland.controller;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ivandanylov.movieland.service.MovieService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class MovieController {
-    private MovieService movieService;
-
-    @RequestMapping(path = "/movie", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public String getAll() throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        return objectMapper.writeValueAsString(movieService.getAll());
-    }
-
-    @RequestMapping(path = "/movie/random", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public String getRandom() throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        return objectMapper.writeValueAsString(movieService.getRandom(3));
-    }
-
-    @Autowired
-    public void setMovieService(MovieService movieService) {
-        this.movieService = movieService;
-    }
-}
