--- src/main/java/com/ivandanylov/movieland/dao/MovieDao.java
@@ -1,10 +1,0 @@
-package com.ivandanylov.movieland.dao;
-
-import com.ivandanylov.movieland.entity.Movie;
-
-import java.util.List;
-
-public interface MovieDao {
-    List<Movie> getAll();
-    List<Movie> getRandom(int count);
-}
