--- src/main/java/com/ivandanylov/movieland/service/MovieService.java
@@ -1,10 +1,0 @@
-package com.ivandanylov.movieland.service;
-
-import com.ivandanylov.movieland.entity.Movie;
-
-import java.util.List;
-
-public interface MovieService {
-    List<Movie> getAll();
-    List<Movie> getRandom(int count);
-}
