--- src/main/resources/properties/database_config.yml
@@ -1,4 +1,0 @@
-application:
-  database:
-    urlEnvironmentVariableName: "JDBC_DATABASE_URL"
-    driverClassName: "org.postgresql.Driver"
