--- src/main/resources/properties/database_query.yml
@@ -1,27 +1,0 @@
-getAllMovies: >-
-  SELECT
-    id,
-    name,
-    original_name,
-    issue_date,
-    rating,
-    price,
-    poster_link
-  FROM
-      movieland.movie;
-
-getRandomMovies: >-
-  SELECT
-    id,
-    name,
-    original_name,
-    issue_date,
-    rating,
-    price,
-    poster_link
-  FROM
-    movieland.movie
-  ORDER BY
-    random(),
-    id
-  LIMIT ?;
