--- src/test/java/com/ivandanylov/movieland/config/yaml/YamlTestMovie.java
@@ -1,20 +1,0 @@
-package com.ivandanylov.movieland.config.yaml;
-
-public class YamlTestMovie {
-    private YamlTestSql movie;
-
-    public YamlTestSql getMovie() {
-        return movie;
-    }
-
-    public void setMovie(YamlTestSql movie) {
-        this.movie = movie;
-    }
-
-    @Override
-    public String toString() {
-        return "YamlTestMovie{" +
-                "movie=" + movie +
-                '}';
-    }
-}
