--- src/test/java/com/ivandanylov/movieland/config/yaml/YamlTestMovieland.java
@@ -1,20 +1,0 @@
-package com.ivandanylov.movieland.config.yaml;
-
-public class YamlTestMovieland {
-    private YamlTestMovie movieland;
-
-    public YamlTestMovie getMovieland() {
-        return movieland;
-    }
-
-    public void setMovieland(YamlTestMovie movieland) {
-        this.movieland = movieland;
-    }
-
-    @Override
-    public String toString() {
-        return "YamlTestMovieland{" +
-                "movieland=" + movieland +
-                '}';
-    }
-}
