--- src/test/resources/hsqldb.yml
@@ -1,18 +1,0 @@
-movieland:
-  movie:
-    createTableSql: >-
-      CREATE TABLE movie(
-        id INT NOT NULL,
-        name VARCHAR(50) NOT NULL,
-        original_name VARCHAR(50) NOT NULL,
-        issue_date DATE NOT NULL,
-        price NUMERIC(8,2) NOT NULL,
-        rating NUMERIC(4,2) NOT NULL,
-        poster_link VARCHAR(50),
-        PRIMARY KEY (id)
-      )
-    dropTableSql: DROP TABLE movie
-    insertSql: >-
-      INSERT INTO movie VALUES (1,'Movie name', 'Original movie name', DATE '1980-04-09', 123.45, 9.8, 'https://link.com/image.jpg')
-    getAllSql: SELECT * FROM movie
-    getRandomSql: getRandomSql
